backend/requirements.txt
# Django Core
django>=4.0,<5.0  # Specify a compatible range

# REST API
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.0,<6.0
django-cors-headers>=4.0,<5.0

# Utilities
python-dotenv>=1.0,<2.0
bcrypt>=4.0,<5.0

# Production Server (Optional for now, uncomment later)
# gunicorn>=21.0,<22.0

# Database (Optional for now, uncomment if using PostgreSQL)
# psycopg2-binary>=2.9,<3.0